+++ src/steps/resource-manager/authorization/client.ts
@@ -4,35 +4,33 @@ import {
   RoleAssignment,
   ClassicAdministrator,
 } from '@azure/arm-authorization/esm/models';
-import { Client } from '../../../azure/resource-manager/client';
+import {
+  Client,
+  iterateAllResources,
+} from '../../../azure/resource-manager/client';
 import { IntegrationProviderAPIError } from '@jupiterone/integration-sdk-core';
 
 export class AuthorizationClient extends Client {
-  public async getRoleDefinition(
-    fullyQualifiedRoleDefinitionId: string,
-  ): Promise<RoleDefinition | undefined> {
+  public async iterateRoleDefinitions(
+    subscriptionId: string,
+    callback: (rd: RoleDefinition) => void | Promise<void>,
+  ): Promise<void> {
     const serviceClient = await this.getAuthenticatedServiceClient(
       AuthorizationManagementClient,
     );
 
-    try {
-      const roleDefinition = await serviceClient.roleDefinitions.getById(
-        fullyQualifiedRoleDefinitionId,
-      );
-      return roleDefinition;
-    } catch (err) {
-      /* istanbul ignore else */
-      if (err.statusCode === 404) {
-        this.logger.warn({ err }, 'Resource not found');
-      } else {
-        throw new IntegrationProviderAPIError({
-          cause: err,
-          endpoint: 'authorization.roleDefinition',
-          status: err.statusCode,
-          statusText: err.statusText,
-        });
-      }
-    }
+    await iterateAllResources({
+      serviceClient,
+      resourceEndpoint: {
+        list: async () => {
+          return serviceClient.roleDefinitions.list(subscriptionId);
+        },
+        listNext: serviceClient.roleDefinitions.listNext,
+      },
+      resourceDescription: 'authorization.roleDefinitions',
+      logger: this.logger,
+      callback,
+    });
   }
 
   public async iterateRoleAssignments(