+++ src/steps/resource-manager/utils/matchers.ts
@@ -3,11 +3,29 @@ export const EOL_MATCHER = '$';
 export const SUBSCRIPTION_MATCHER = '/subscriptions/([^/]+)';
 export const RESOURCE_GROUP_MATCHER =
   SUBSCRIPTION_MATCHER + '/resource[G|g]roups/([^/]+)';
+
+/**
+ * Returns true if `azureResourceId` contains valid resource group
+ */
+function containsSubscriptionId(azureResourceId: string): boolean {
+  return new RegExp(SUBSCRIPTION_MATCHER, 'i').test(azureResourceId);
+}
+
+export function getSubscriptionId(azureResourceId: string | undefined) {
+  if (!azureResourceId) return;
+
+  if (containsSubscriptionId(azureResourceId)) {
+    return (azureResourceId.match(
+      new RegExp(SUBSCRIPTION_MATCHER, 'i'),
+    ) as RegExpMatchArray)[0];
+  }
+}
+
 /**
  * Returns true if `azureResourceId` contains valid resource group
  */
 function containsResourceGroup(azureResourceId: string): boolean {
-  return new RegExp(RESOURCE_GROUP_MATCHER).test(azureResourceId);
+  return new RegExp(RESOURCE_GROUP_MATCHER, 'i').test(azureResourceId);
 }
 
 /**
@@ -19,7 +37,7 @@ function getNonNormalizedResourceGroupId(
 ): string | undefined {
   if (containsResourceGroup(azureResourceId)) {
     return (azureResourceId.match(
-      new RegExp(RESOURCE_GROUP_MATCHER),
+      new RegExp(RESOURCE_GROUP_MATCHER, 'i'),
     ) as RegExpMatchArray)[0];
   }
 }